requirements.txt
pandas
numpy
python-binance
flask
flask-script
flask-restplus
seaborn